% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sodaR.R
\name{posixify}
\alias{posixify}
\title{Convert Socrata calendar_date string to POSIX}
\usage{
posixify(x)
}
\arguments{
\item{x}{- A character vector containing date-time strings in various formats that will be converted to a POSIX object}
}
\value{
a POSIX date
}
\description{
Convert Socrata calendar_date string to POSIX
}
\author{
Hugh J. Devlin, Ph. D. \email{Hugh.Devlin@cityofchicago.org}
}
